finished cart-items

+++ src/cart-item/cart.service.ts
@@ -1,7 +1,6 @@
 import { Injectable, NotFoundException } from '@nestjs/common';
 import { InjectRepository } from '@nestjs/typeorm';
 import { Repository } from 'typeorm';
-// import { CartItem } from 'src/entities/cart-item.entity';
 import { Cart } from 'src/entities/cart.entity';
 import { AuthService } from 'src/auth/auth.service';
 import { ProductsService } from 'src/products/products.service';
@@ -29,26 +28,18 @@ export class CartService {
       });
     }
 
-    async removeFromCart(cartItemId: number): Promise<void> {
-
-      // Проверяем, существует ли корзина пользователя
-      // const user = await this.authService.findOne(userId);
-      // if (!user) {
-      //   throw new NotFoundException(`User with ID ${userId} not found`);
-      // }
-  
-      // // Находим элемент корзины пользователя
-      // const cartItem = user.cartItems.find(item => item.id === cartItemId);
-      // if (!cartItem) {
-      //   throw new NotFoundException(`CartItem with ID ${cartItemId} not found in the user's cart`);
-      // }
-  
-      // Удаляем элемент из корзины
-      const cartItem = await this.cartRepository.findOne({where:{id:cartItemId}})
-      await this.cartRepository.delete({ id: cartItemId });
-;
+    async removeAllItemsForUser(userId: number): Promise<void> {
+      await this.cartRepository.delete({ user: { id: userId } });
     }
     
+
+    async removeFromCart(cartItemId: number): Promise<void> {
+      const cartItem = await this.cartRepository.findOne({where:{id:cartItemId}})
+      if (!cartItem){
+        throw new NotFoundException(`CartItem with ID ${cartItemId} not found in the user's cart`);
+      }
+      await this.cartRepository.delete({ id: cartItemId });
+    }
   }
